+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ Unreleased
 Added
 -----
 - Python 3.10 support.
+- Option to pass figure initialization keyword arguments to Matplotlib via plotting
+  methods.
 
 Fixed
 -----